+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpLogExporterTests.cs
@@ -17,6 +17,8 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation;
 using OpenTelemetry.Internal;
@@ -30,6 +32,87 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 {
     public class OtlpLogExporterTests : Http2UnencryptedSupportTests
     {
+        [Fact]
+        public void AddOtlpLogExporterOptionstest()
+        {
+            AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);
+            var loggerOptions = new OpenTelemetryLoggerOptions();
+            Assert.False(loggerOptions.ParseStateValues);
+            loggerOptions.AddOtlpExporter();
+            Assert.True(loggerOptions.ParseStateValues);
+        }
+
+        [Fact]
+        public void AddOtlpLogExporterSetsParseStateValueToTrue()
+        {
+            AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);
+            var logRecords = new List<LogRecord>();
+            using var loggerFactory = LoggerFactory.Create(builder =>
+            {
+                builder.AddOpenTelemetry(options =>
+                {
+                    options.AddInMemoryExporter(logRecords);
+                    options.AddOtlpExporter();
+                });
+            });
+
+            var logger = loggerFactory.CreateLogger("OtlpLogExporterTests");
+            logger.LogInformation("Hello from {name} {price}.", "tomato", 2.99);
+            Assert.Single(logRecords);
+            var logRecord = logRecords[0];
+            Assert.Null(logRecord.State);
+            Assert.NotNull(logRecord.StateValues);
+        }
+
+        [Fact]
+        public void AddOtlpLogExporterParseStateValueCanBeTurnedOff()
+        {
+            AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);
+            var logRecords = new List<LogRecord>();
+            using var loggerFactory = LoggerFactory.Create(builder =>
+            {
+                builder.AddOpenTelemetry(options =>
+                {
+                    options.AddInMemoryExporter(logRecords);
+                    options.AddOtlpExporter();
+                    options.ParseStateValues = false;
+                });
+            });
+
+            var logger = loggerFactory.CreateLogger("OtlpLogExporterTests");
+            logger.LogInformation("Hello from {name} {price}.", "tomato", 2.99);
+            Assert.Single(logRecords);
+            var logRecord = logRecords[0];
+            Assert.NotNull(logRecord.State);
+            Assert.Null(logRecord.StateValues);
+        }
+
+        [Fact]
+        public void AddOtlpLogExporterParseStateValueCanBeTurnedOffHosting()
+        {
+            var logRecords = new List<LogRecord>();
+
+            AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);
+            var hostBuilder = new HostBuilder();
+            hostBuilder.ConfigureLogging(logging => logging.AddOpenTelemetry(options =>
+            {
+                options.AddInMemoryExporter(logRecords);
+                options.AddOtlpExporter();
+            }));
+
+            hostBuilder.ConfigureServices(services =>
+            services.Configure<OpenTelemetryLoggerOptions>(options => options.ParseStateValues = false));
+
+            var host = hostBuilder.Build();
+            var loggerFactory = host.Services.GetService<ILoggerFactory>();
+            var logger = loggerFactory.CreateLogger("OtlpLogExporterTests");
+            logger.LogInformation("Hello from {name} {price}.", "tomato", 2.99);
+            Assert.Single(logRecords);
+            var logRecord = logRecords[0];
+            Assert.NotNull(logRecord.State);
+            Assert.Null(logRecord.StateValues);
+        }
+
         [Fact]
         public void OtlpLogRecordTestWhenStateValuesArePopulated()
         {
@@ -225,8 +308,10 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             }
         }
 
-        [Fact]
-        public void CheckToOtlpLogRecordFormattedMessage()
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public void CheckToOtlpLogRecordBodyIsPopulated(bool includeFormattedMessage)
         {
             var logRecords = new List<LogRecord>();
             using var loggerFactory = LoggerFactory.Create(builder =>
@@ -234,11 +319,14 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
                 builder.AddOpenTelemetry(options =>
                 {
                     options.AddInMemoryExporter(logRecords);
-                    options.IncludeFormattedMessage = true;
+                    options.IncludeFormattedMessage = includeFormattedMessage;
+                    options.ParseStateValues = true;
                 });
             });
 
             var logger = loggerFactory.CreateLogger("OtlpLogExporterTests");
+
+            // Scenario 1 - Using ExtensionMethods on ILogger.Log
             logger.LogInformation("OpenTelemetry {Greeting} {Subject}!", "Hello", "World");
             Assert.Single(logRecords);
 
@@ -246,7 +334,56 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             var otlpLogRecord = logRecord.ToOtlpLog();
 
             Assert.NotNull(otlpLogRecord);
-            Assert.Equal(logRecord.FormattedMessage, otlpLogRecord.Body.StringValue);
+            if (includeFormattedMessage)
+            {
+                Assert.Equal(logRecord.FormattedMessage, otlpLogRecord.Body.StringValue);
+            }
+            else
+            {
+                Assert.Equal("OpenTelemetry {Greeting} {Subject}!", otlpLogRecord.Body.StringValue);
+            }
+
+            logRecords.Clear();
+
+            // Scenario 2 - Using the raw ILogger.Log Method
+            logger.Log(LogLevel.Information, default, "state", exception: null, (st, ex) => "Formatted Message");
+            Assert.Single(logRecords);
+
+            logRecord = logRecords[0];
+            otlpLogRecord = logRecord.ToOtlpLog();
+
+            Assert.NotNull(otlpLogRecord);
+            if (includeFormattedMessage)
+            {
+                Assert.Equal(logRecord.FormattedMessage, otlpLogRecord.Body.StringValue);
+            }
+            else
+            {
+                Assert.Null(otlpLogRecord.Body);
+            }
+
+            logRecords.Clear();
+
+            // Scenario 3 - Using the raw ILogger.Log Method, but with null
+            // formatter.
+            logger.Log(LogLevel.Information, default, "state", exception: null, formatter: null);
+            Assert.Single(logRecords);
+
+            logRecord = logRecords[0];
+            otlpLogRecord = logRecord.ToOtlpLog();
+
+            Assert.NotNull(otlpLogRecord);
+
+            // There is no formatter, so no way to populate Body.
+            // Exporter won't even attempt to do ToString() on State.
+            if (includeFormattedMessage)
+            {
+                Assert.Null(otlpLogRecord.Body);
+            }
+            else
+            {
+                Assert.Null(otlpLogRecord.Body);
+            }
         }
 
         [Fact]