+++ test/OpenTelemetry.Tests/Logs/LoggerFactoryAndResourceBuilderTests.cs
@@ -13,7 +13,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
-#if !NET461
 
 using System;
 using System.Collections.Generic;
@@ -79,4 +78,3 @@ namespace OpenTelemetry.Logs.Tests
         }
     }
 }
-#endif